+++ utils/ft_color_utils.c
@@ -6,7 +6,7 @@
 /*   By: rmeiboom <rmeiboom@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2021/04/14 13:46:56 by rmeiboom      #+#    #+#                 */
-/*   Updated: 2021/04/14 14:02:03 by rmeiboom      ########   odam.nl         */
+/*   Updated: 2021/04/21 20:00:01 by rmeiboom      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ void	color_check(t_rgb *col)
 		col->b = 0;
 }
 
-void	color_multi(t_rgb *col, float factor)
+void	color_multi(t_rgb *col, double factor)
 {
 	col->r *= factor;
 	col->g *= factor;
