+++ utils/vec_utils.c
@@ -6,13 +6,13 @@
 /*   By: rmeiboom <marvin@codam.nl>                   +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2021/03/07 20:49:30 by rmeiboom      #+#    #+#                 */
-/*   Updated: 2021/04/07 17:17:22 by rmeiboom      ########   odam.nl         */
+/*   Updated: 2021/04/21 19:50:31 by rmeiboom      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minirt.h"
 
-float	vec_len(t_vec vec)
+double	vec_len(t_vec vec)
 {
 	return(sqrt(pow((vec.x), 2) + pow((vec.y), 2) + pow((vec.z), 2)));	
 }
@@ -35,7 +35,7 @@ t_vec	vec_minus(t_vec a, t_vec b)
 	return (a);
 }
 
-t_vec	vec_multiply(t_vec a, float factor)
+t_vec	vec_multiply(t_vec a, double factor)
 {
 	a.x *= factor;
 	a.y *= factor;
@@ -44,7 +44,7 @@ t_vec	vec_multiply(t_vec a, float factor)
 	return (a);
 }
 
-t_vec	vec_divide(t_vec a, float factor)
+t_vec	vec_divide(t_vec a, double factor)
 {
 	a.x /= factor;
 	a.y /= factor;
